+++ packages/spacecat-shared-project-engine-client/CHANGELOG.md
@@ -1,3 +1,21 @@
+## [@adobe/spacecat-shared-project-engine-client-v1.23.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-project-engine-client-v1.22.0...@adobe/spacecat-shared-project-engine-client-v1.23.0) (2026-09-09)
+
+### Features
+
+* **project-engine-client:** add BCP-47 code to the language catalog ([#1919](https://github.com/adobe/spacecat-shared/issues/1919)) ([4092604](https://github.com/adobe/spacecat-shared/commit/4092604f00a5642e91b690f73ec2acb6c26352a4))
+
+## [@adobe/spacecat-shared-project-engine-client-v1.22.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-project-engine-client-v1.21.1...@adobe/spacecat-shared-project-engine-client-v1.22.0) (2026-09-08)
+
+### Features
+
+* **project-engine-mock:** add native nested tag fixtures ([#1916](https://github.com/adobe/spacecat-shared/issues/1916)) ([591e183](https://github.com/adobe/spacecat-shared/commit/591e1838e2226d30b252737f9dd8c099dd93321a)), closes [adobe/serenity-docs#26](https://github.com/adobe/serenity-docs/issues/26)
+
+## [@adobe/spacecat-shared-project-engine-client-v1.21.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-project-engine-client-v1.21.0...@adobe/spacecat-shared-project-engine-client-v1.21.1) (2026-09-07)
+
+### Bug Fixes
+
+* **project-engine-client:** honour main_brand:true on v2 benchmark create (LLMO-7421) ([#1918](https://github.com/adobe/spacecat-shared/issues/1918)) ([57f37d1](https://github.com/adobe/spacecat-shared/commit/57f37d1f4e4561119d9148671dac09f2259f0fcb))
+
 ## [@adobe/spacecat-shared-project-engine-client-v1.21.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-project-engine-client-v1.20.0...@adobe/spacecat-shared-project-engine-client-v1.21.0) (2026-08-31)
 
 ### Features